requirements.txt
# Core Backend
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-dotenv==1.0.0
pydantic==2.5.3

# LLM Providers
google-generativeai==0.4.0
httpx==0.26.0  # For Groq API calls

# Speech Processing
openai-whisper==20231117
faster-whisper==1.0.3  # Optimized Whisper with CTranslate2
webrtcvad==2.0.10  # Voice Activity Detection
soundfile==0.12.1
numpy==1.26.3

# Edge TTS for free, high-quality voice synthesis
edge-tts==6.1.9

# HTTP & Async
python-multipart==0.0.6

# Utilities
uuid6==2024.1.12

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
